Move display aggregations into their own namespace
