+++ search/src/main/scala/weco/api/search/models/display/DisplayImageAggregations.scala
@@ -1,13 +1,14 @@
-package weco.api.search.models
+package weco.api.search.models.display
 
-import io.circe.generic.extras.semiauto._
 import io.circe.Encoder
 import io.circe.generic.extras.JsonKey
+import io.circe.generic.extras.semiauto._
+import weco.api.search.models
+import weco.api.search.models.{Aggregation, ImageAggregations}
 import weco.catalogue.display_model.locations.DisplayLicense
 import weco.catalogue.display_model.work.{DisplayAbstractAgent, DisplayGenre}
 import weco.catalogue.internal_model.identifiers.IdState.Minted
 import weco.catalogue.internal_model.work.{AbstractAgent, Genre}
-import weco.http.json.DisplayJsonUtil._
 
 case class DisplayImageAggregations(
   license: Option[DisplayAggregation[DisplayLicense]],
@@ -19,7 +20,6 @@ case class DisplayImageAggregations(
 )
 
 object DisplayImageAggregations {
-  import weco.catalogue.display_model.Implicits._
 
   implicit def encoder: Encoder[DisplayImageAggregations] =
     deriveConfiguredEncoder
@@ -43,6 +43,6 @@ object DisplayImageAggregations {
     display: T => D
   ): Option[DisplayAggregation[D]] =
     maybeAgg.map {
-      DisplayAggregation(_, display)
+      models.display.DisplayAggregation(_, display)
     }
 }
